advent: display leaderboards as a colour image
uses pillow to render leaderboards into an image with a style
very close to the website. inline display in the discord app
is a bit smaller than i would've liked but alas.

text format is kept alongside the image since it's probably nice
to have. maybe this should be behind an argument, but there are
already so many.

implementation has some weaknesses:
- unicode multi-character sequences and emoji aren't rendered
- relies very much on a monospace font for alignment

let me know if this is something you want.