starboard: support starring system messages
including system messages originating inside starboard.

this changes the lookup logic to attempt both kinds of lookups
(both received message and starboard message), allowing for
starboard messages to be mixed with original messages in the starboard
channel.

this also changes the referenced message handling to attempt to fetch
the referenced message if it has not been resolved. this gets us the
full pinned message for pins_add events.

also tweaks the display of referenced messages to handle kinds of
reference which are not a reply.

in delete, we also have to use `db_session.delete(entry)` instead of
`entry.delete()`. i do not know why - renamed method after update, maybe?